feat: changed the meetup request url to a beetr one
As far as I could check, the following url is the best way to ask for
next or last meetup:

https://api.meetup.com/montrealelixir/events\?scroll=future_or_past&page=1

The page=1 param makes sure there is always only one result returned.
The scroll=future_or_past makes it return the next upcoming event if
available, otherwise it returns the last event.